fix(sync): auto-resume bootstrap after auth halt + health sync probe
First-import no longer stays halted on transient token failures: status
poll clears the breaker and reschedules. /health exposes sync progress
for ops checks without a session.
