+++ worker/src/sync/sync.ts
@@ -36,6 +36,9 @@ import {
 import { closedHopPass, edgesForRepo, flushEdges } from "./edges";
 import type { EdgeData } from "./label-vocab";
 import { discoverTenants } from "./tenants";
+import { type RunSyncOptions, selectWindowedRepos } from "./window";
+
+export type { RunSyncOptions } from "./window";
 
 // ---------------------------------------------------------------------------
 // Public facade — re-export the split-out symbols so existing importers
@@ -61,23 +64,6 @@ export { syncRepoIssues } from "./repo-issues";
 export { syncBranches } from "./repo-branches";
 export { syncRepoBundle } from "./bundle";
 export { writeRunAudit } from "./audit";
-
-export interface RunSyncOptions {
-  /** Bootstrap: sync only repos missing sync_state (up to WINDOW), skip slot rotation. */
-  prioritizeUnsynced?: boolean;
-}
-
-async function listUnsyncedRepos(db: D1Database, allRepos: string[]): Promise<string[]> {
-  const rows = await db
-    .prepare(
-      `SELECT s.repo FROM sync_state s
-       INNER JOIN repos r ON r.repo = s.repo
-       WHERE s.last_synced_at IS NOT NULL AND TRIM(s.last_synced_at) != ''`,
-    )
-    .all<{ repo: string }>();
-  const synced = new Set((rows.results ?? []).map((r) => r.repo));
-  return allRepos.filter((r) => !synced.has(r));
-}
 
 /**
  * Org-wide sync: check halt → acquire lock → enumerate repos → two-pass issue
@@ -234,32 +220,14 @@ export async function runSync(env: Env, opts?: RunSyncOptions): Promise<void> {
       console.log(`[sync] pruned data for ${staleReposPruned} stale repo(s)`);
     }
 
-    // Phase 2 — repo fan-out. Bootstrap prioritises unsynced repos so we do not
-    // waste a pass re-fetching repos that already have sync_state (26/39 stall).
-    let windowedRepos: string[];
-    let slot = 0;
-    if (opts?.prioritizeUnsynced) {
-      const unsynced = await listUnsyncedRepos(db, allRepos);
-      windowedRepos = unsynced.slice(0, WINDOW);
-      console.log(
-        `[sync] bootstrap unsynced=${unsynced.length} syncing=${windowedRepos.length}/${allRepos.length}`,
-      );
-      if (windowedRepos.length === 0) {
-        outcome = "success";
-        return;
-      }
-    } else {
-      const slotRow = await db
-        .prepare(`SELECT value FROM sync_control WHERE key='sync_slot' AND tenant_id = 0`)
-        .first<{ value: string }>();
-      slot = Number.parseInt(slotRow?.value ?? "0", 10);
-      const windowStart = slot * WINDOW;
-      const windowEnd = windowStart + WINDOW;
-      // Windowing only engages past WINDOW repos; below that, sync everything hourly.
-      windowedRepos = allRepos.length <= WINDOW ? allRepos : allRepos.slice(windowStart, windowEnd);
-      console.log(
-        `[sync] slot=${slot} window=[${windowStart},${windowEnd}) repos=${windowedRepos.length}/${allRepos.length}`,
-      );
+    const {
+      windowedRepos,
+      slot,
+      empty: noReposToSync,
+    } = await selectWindowedRepos(db, allRepos, opts);
+    if (noReposToSync) {
+      outcome = "success";
+      return;
     }
 
     // Per-repo token resolver: try owning tenant first, fall back down list.
